multiple-agents-system/pyproject.toml
[project]
name = "multiple-agents-system"
version = "0.1.0"
description = "AWS AgentCore Production Multi-Agent System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "langchain>=0.1.0",
    "langchain-aws>=0.1.0",
    "langgraph>=0.1.0",
    "structlog>=24.1.0",
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.26.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    # Local development
    "anthropic>=0.40.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "config", "lambda"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
